Bug 1396797 Support for partials in manifest
The outbound manifest, used by balrog scriptworker, needs to
have information about partials added.

It's unwise to add a partial if there's no related complete,
so this involves a refactor so that instead of directly
populating context.balrog_manifest, we instead fill a
staging area with details about the files we have uploaded.

Once all uploads are done, we generate the balrog manifest
to be of the expected form.